Complete the registration flow.
This moves the crypto deps to the common package as that is where
we're calculating the hash now. It also adds the functionality to
register a user from RPC, returning the User entity to the RPC server
and the ApiKey to the client.

We need a better mechanism of generating the apikey (it's just ahas of
the email right now) as well as a better way of returning it (should
be emailed to them) but this unblocks progress for now.